src/apps/nfc/summary/controllers/insert-summaries.controller.spec.ts
import { Test, TestingModule } from '@nestjs/testing';

// custom items
import { InsertSummariesController } from './insert-summaries.controller';
import { ICommandBus } from '@hades/shared/domain/bus/command-bus';
import { IQueryBus } from '@hades/shared/domain/bus/query-bus';
import { summaries } from '@hades/nfc/summary/infrastructure/seeds/summary.seed'

describe('InsertSummariesController', () => 
{
    let controller: InsertSummariesController;
    let queryBus: IQueryBus;
    let commandBus: ICommandBus;

    beforeEach(async () => 
    {
        const module: TestingModule = await Test.createTestingModule({
            controllers: [
                InsertSummariesController
            ],
            providers: [
                {
                    provide: IQueryBus,
                    useValue: {
                        ask: () => {},
                    }
                },
                {
                    provide: ICommandBus,
                    useValue: {
                        dispatch: () => {},
                    }
                },
            ]
        }).compile();

        controller  = module.get<InsertSummariesController>(InsertSummariesController);
        queryBus    = module.get<IQueryBus>(IQueryBus);
        commandBus  = module.get<ICommandBus>(ICommandBus);
    });

    it('InsertSummariesController should be defined', () => 
    {
        expect(controller).toBeDefined();
    });

    // Test get method
    describe('main', () => 
    {
        it('InsertSummariesController should be defined', () => 
        {
            expect(controller).toBeDefined();
        });

        it('should return an summaries created', async () => 
        {
            expect(await controller.main(summaries)).toBe(undefined);
        });
    });
});